requirements.txt
# Core Framework
Flask>=2.0.0
Werkzeug>=2.0.0

# AI and ML
google-generativeai>=0.3.0
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.1
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
Pillow>=9.0.0
matplotlib>=3.5.0

# Medical Image Support
pydicom>=2.3.0

# Text Processing
python-dotenv>=0.19.0
gTTS>=2.3.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

